requirements.txt
oci==2.99.1
python-dotenv==1.0.0
